pyproject.toml
[project]
name = "HakuBot"
version = "0.1.0"
description = "HakuBot"
readme = "README.md"
requires-python = ">=3.10, <4.0"
dependencies = [
    "nonebot2[fastapi]>=2.4.4",
    "nonebot-adapter-onebot>=2.4.6",
    # NoneBot 插件
    "nonebot-plugin-apscheduler",
    "nonebot-plugin-htmlrender",
    "nonebot-plugin-localstore",
    "nonebot-plugin-send-anything-anywhere",
    "nonebot-plugin-session",
    "nonebot-plugin-alconna",
    "nonebot-plugin-userinfo",
    # 通用依赖
    "httpx",
    "aiohttp",
    "aiofiles",
    "pillow",
    "pil-utils",
    "numpy",
    "opencv-python-headless",
    "beautifulsoup4",
    "pytz",
    "curl-cffi",
    "pydub",
    "aiosqlite",
    "tenacity",
    "jinja2",
    "emoji",
    "playwright",
    "sqlmodel",
    "PicImageSearch",
    "pydantic",
    "pilmoji>=2.0.5",
    "colour>=0.1.5",
    "msgspec>=0.20.0",
    "watchdog>=6.0.0",
    "sssekai>=0.7.62",
    "nonebot-bison>=0.9.14",
    "json-repair>=0.58.0",
    "psutil>=7.2.2",
]

[tool.nonebot]
plugin_dirs = ["plugins"]
builtin_plugins = []

[tool.nonebot.adapters]
nonebot-adapter-onebot = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
"@local" = []

[tool.nonebot.plugins]
plugins = ["nonebot_bison"]
